Removed code used for serve drive and started work on replacement code for new drive.
