+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TestAutonomous.java
@@ -1,11 +1,10 @@
 package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
 //@Disabled
 @Autonomous(name="testing", group="testing")
-public class TestAutonomous extends FunctionLibrary {
+public class TestAutonomous extends Automation {
 
     @Override
     public void runOpMode() throws InterruptedException {
